+++ Makefile
@@ -34,8 +34,7 @@ fix:
 
 .PHONY: test
 test:
-	cargo test --features vpicc,rsa4096-gen gpg_crypto,sequoia_gen_key
-	cargo test --features vpicc,rsa4096
+	cargo test --features vpicc,rsa4096-gen 
 	
 
 .PHONY: test
@@ -43,7 +42,7 @@ dangerous-real-card-test:
 	ps aux | grep pcscd | grep -v grep || sudo pcscd
 	cargo test --features rsa2048-gen,rsa4096,dangerous-test-real-card sequoia
 	sudo pkill pcscd
-	cargo test --features rsa2048-gen,rsa4096,dangerous-test-real-card gpg
+	cargo test --features rsa2048-gen,rsa4096-gen,dangerous-test-real-card _hardware
 
 .PHONY: fuzz
 fuzz: fuzz-corpus
